webapp/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
flower==1.0.0
humanize==3.12.0
idna==3.2
kombu==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
